[action/deploy/environment] rename from action/deploy
This was always the deploy environment action but was initially
collapsed under a lower namespace. The filename now respects the class
namespacing.
